rename lib file api_formatting.py into dku_amazon_comprehend_medical.py
